--- src/main/java/ladysnake/gaspunk/client/particle/ParticleManager.java
@@ -1,119 +1,0 @@
-package ladysnake.gaspunk.client.particle;
-
-import ladysnake.gaspunk.GasPunkConfig;
-import ladysnake.gaspunk.api.IGas;
-import ladysnake.gaspunk.api.IGasParticleType;
-import ladysnake.gaspunk.init.ModGases;
-import net.minecraft.client.Minecraft;
-import net.minecraft.client.particle.Particle;
-import net.minecraft.client.renderer.ActiveRenderInfo;
-import net.minecraft.client.renderer.BufferBuilder;
-import net.minecraft.client.renderer.GlStateManager;
-import net.minecraft.client.renderer.Tessellator;
-import net.minecraft.client.renderer.texture.TextureMap;
-import net.minecraft.client.renderer.vertex.DefaultVertexFormats;
-import net.minecraft.entity.player.EntityPlayer;
-import net.minecraftforge.client.event.RenderWorldLastEvent;
-import net.minecraftforge.client.event.TextureStitchEvent;
-import net.minecraftforge.fml.common.Mod;
-import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
-import net.minecraftforge.fml.common.gameevent.TickEvent;
-import net.minecraftforge.fml.relauncher.Side;
-import net.minecraftforge.fml.relauncher.SideOnly;
-import org.lwjgl.opengl.GL11;
-
-import java.util.HashSet;
-import java.util.Set;
-
-/**
- * This class has been adapted from embers' source code under GNU Lesser General Public License 2.1
- * https://github.com/RootsTeam/Embers/blob/master/src/main/java/teamroots/embers/particle/ParticleRenderer.java
- *
- * @author Elucent
- */
-@SideOnly(Side.CLIENT)
-@Mod.EventBusSubscriber(Side.CLIENT)
-public class ParticleManager {
-
-    public static final ParticleManager INSTANCE = new ParticleManager();
-
-    private final Set<Particle> particles = new HashSet<>();
-
-    @SubscribeEvent
-    public static void onTextureStitch(TextureStitchEvent.Pre event) {
-        // gets every possible particle texture from registered gases and register them as sprites
-        ModGases.REGISTRY.getValues().stream()
-                .map(IGas::getParticleType)
-                .map(IGasParticleType::getParticleTexture)
-                .distinct()
-                .forEach(event.getMap()::registerSprite);
-    }
-
-    @SubscribeEvent
-    public static void onGameTick(TickEvent.ClientTickEvent event) {
-        INSTANCE.updateParticles();
-    }
-
-    @SubscribeEvent
-    public static void onRenderWorldLast(RenderWorldLastEvent event) {
-        INSTANCE.renderParticles(event.getPartialTicks());
-    }
-
-    public void updateParticles() {
-        // particles cost a lot less to update than to render
-        particles.stream().limit(3* GasPunkConfig.client.maxParticles).forEach(Particle::onUpdate);
-        particles.removeIf(p -> !p.isAlive());
-    }
-
-    public void renderParticles(float partialTicks) {
-        float f = ActiveRenderInfo.getRotationX();
-        float f1 = ActiveRenderInfo.getRotationZ();
-        float f2 = ActiveRenderInfo.getRotationYZ();
-        float f3 = ActiveRenderInfo.getRotationXY();
-        float f4 = ActiveRenderInfo.getRotationXZ();
-        EntityPlayer player = Minecraft.getMinecraft().player;
-        if (player != null) {
-            Particle.interpPosX = player.lastTickPosX + (player.posX - player.lastTickPosX) * partialTicks;
-            Particle.interpPosY = player.lastTickPosY + (player.posY - player.lastTickPosY) * partialTicks;
-            Particle.interpPosZ = player.lastTickPosZ + (player.posZ - player.lastTickPosZ) * partialTicks;
-            Particle.cameraViewDir = player.getLook(partialTicks);
-
-            GlStateManager.pushMatrix();
-
-            GlStateManager.enableAlpha();
-            GlStateManager.enableBlend();
-            GlStateManager.alphaFunc(516, 0.003921569F);
-            GlStateManager.disableCull();
-
-            GlStateManager.depthMask(false);
-
-            Minecraft.getMinecraft().renderEngine.bindTexture(TextureMap.LOCATION_BLOCKS_TEXTURE);
-            Tessellator tess = Tessellator.getInstance();
-            BufferBuilder buffer = tess.getBuffer();
-
-            // render normal particles
-            {
-                GlStateManager.blendFunc(GlStateManager.SourceFactor.SRC_ALPHA, GlStateManager.DestFactor.ONE_MINUS_SRC_ALPHA);
-                buffer.begin(GL11.GL_QUADS, DefaultVertexFormats.PARTICLE_POSITION_TEX_COLOR_LMAP);
-                particles.stream()
-                        .limit(GasPunkConfig.client.maxParticles)
-                        .forEach(p -> p.renderParticle(buffer, player, partialTicks, f, f4, f1, f2, f3));
-                tess.draw();
-            }
-
-            GlStateManager.enableCull();
-            GlStateManager.depthMask(true);
-            GlStateManager.blendFunc(GlStateManager.SourceFactor.SRC_ALPHA, GlStateManager.DestFactor.ONE_MINUS_SRC_ALPHA);
-            GlStateManager.disableBlend();
-            GlStateManager.alphaFunc(516, 0.1F);
-
-            GlStateManager.popMatrix();
-        }
-    }
-
-    public void addParticle(Particle p) {
-//        if (particles.size() < GasPunkConfig.client.maxParticles)
-            particles.add(p);
-    }
-
-}
